+++ src/Seek.Cli/Commands.Search.cs
@@ -14,7 +14,8 @@ internal static partial class Commands {
 	/// <param name="regex">-r, Treat the query as regex pattern</param>
 	/// <param name="caseSensitive">Perform a case sensitive search</param>
 	/// <param name="plain">-p, Disable matching section highlight</param>
-	/// <param name="null">Emit machine-readable NUL-terminated plain paths for safe piping</param>
+	/// <param name="absolute">Emit absolute paths instead of paths relative to the selected root</param>
+	/// <param name="null">Emit machine-readable NUL-terminated paths for safe piping, implies --plain and --absolute</param>
 	/// <param name="hidden">-h, Include hidden files and folders</param>
 	/// <param name="system">-s, Include system files and folders</param>
 	/// <param name="files">-f, Match only against files</param>
@@ -28,6 +29,7 @@ internal static partial class Commands {
 		bool regex,
 		bool caseSensitive,
 		bool plain,
+		bool absolute,
 		bool @null,
 		bool hidden,
 		bool system,
@@ -36,64 +38,59 @@ internal static partial class Commands {
 		string root = ".",
 		ConsoleColor highlightColor = ConsoleColor.Green,
 		CancellationToken cancellationToken = default) {
-		SearchType type = (files, directories) switch {
-			(true, false) => SearchType.Files,
-			(false, true) => SearchType.Directories,
-			_ => SearchType.FilesAndDirectories
-		};
+		var search = CreateFileSystemSearch(
+			query,
+			regex,
+			caseSensitive,
+			hidden,
+			system,
+			files,
+			directories,
+			root,
+			cancellationToken);
 
-		var search = new FileSystemSearch(new SearchOptions {
-			Query = query,
-			Root = root,
-			Regex = regex,
-			CaseSensitive = caseSensitive,
-			AttributesToSkip = GetAttributesToSkip(hidden, system),
-			SearchType = type
-		});
-
-		Action<SearchMatch, ConsoleColor> outputHandler = (@null, plain) switch {
+		Action<SearchMatch, bool, ConsoleColor> outputHandler = (@null, plain) switch {
 			(true, _) => WritePlainNullTerminated,
 			(false, true) => WritePlain,
 			_ => WriteRegular
 		};
 
-		await foreach (var searchMatch in search.SearchAsync(cancellationToken).ConfigureAwait(false)) {
-			outputHandler(searchMatch, highlightColor);
+		await foreach (var searchMatch in search.SearchAsync().ConfigureAwait(false)) {
+			outputHandler(searchMatch, absolute, highlightColor);
 		}
 
 		return 0;
 	}
 
-	private static void WriteRegular(SearchMatch searchMatch, ConsoleColor highlightColor) {
-		ReadOnlySpan<char> path = searchMatch.Path;
+	private static void WriteRegular(SearchMatch searchMatch, bool absolute, ConsoleColor highlightColor) {
+		ReadOnlySpan<char> path; int offset;
+		if (!absolute) {
+			path = searchMatch.Path.AsSpan(searchMatch.RelativePathOffset);
+			offset = 0;
+		} else {
+			path = searchMatch.Path.AsSpan();
+			offset = searchMatch.RelativePathOffset;
+			var slice = path.Slice(0, offset);
+			Console.Write(slice, OutputPipe.Out);
+		}
 
 		foreach (var range in searchMatch.Sections) {
 			ConsoleColor color = range.IsMatch ? highlightColor : ConsoleColor.DefaultForeground;
-			var slice = path.Slice(range.Start, range.Length);
+			var slice = path.Slice(offset + range.Start, range.Length);
 			Console.WriteInterpolated($"{color}{slice}");
 		}
 		Console.NewLine();
 	}
 
-	private static void WritePlain(SearchMatch searchMatch, ConsoleColor highlightColor) {
-		Console.WriteLine(searchMatch.Path.AsSpan(), OutputPipe.Out);
+	private static void WritePlain(SearchMatch searchMatch, bool absolute, ConsoleColor highlightColor) {
+		var path = !absolute ? searchMatch.Path.AsSpan(searchMatch.RelativePathOffset) : searchMatch.Path.AsSpan();
+
+		Console.WriteLine(path, OutputPipe.Out);
 	}
 
-	private static void WritePlainNullTerminated(SearchMatch searchMatch, ConsoleColor highlightColor) {
-		Console.WriteInterpolated($"{searchMatch.Path}\0");
-	}
+	private static void WritePlainNullTerminated(SearchMatch searchMatch, bool absolute, ConsoleColor highlightColor) {
+		var path = searchMatch.Path.AsSpan();
 
-	private static FileAttributes GetAttributesToSkip(bool hidden, bool system) {
-		FileAttributes attributesToSkip = FileAttributes.ReparsePoint;
-
-		if (!hidden) {
-			attributesToSkip |= FileAttributes.Hidden;
-		}
-
-		if (!system) {
-			attributesToSkip |= FileAttributes.System;
-		}
-
-		return attributesToSkip;
+		Console.WriteInterpolated($"{path}\0");
 	}
 }